backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
pydantic>=2.7.0
python-dotenv>=1.0.0

# NLP / ML
transformers>=4.40.0
torch>=2.2.0
scikit-learn>=1.4.0
numpy>=1.26.0

# Social APIs
tweepy>=4.14.0
praw>=7.7.0